update readme so new reader can instantly get it
The examples are from the [glom tutorial](https://glom.readthedocs.io/en/latest/tutorial.html) which is just a click away but honestly, the term "path-based access" and the term "spec" was not clear to me as a relatively new Pythonista. Hopefully, including this example helps new readers to understand the purpose of `glom` as quick as possible. Thanks :)